requirements.txt
yfinance==0.2.36
fredapi==0.5.1
duckdb==0.10.0
pandas==2.2.0
exchange-calendars==4.5.4
requests==2.31.0
python-dotenv==1.0.0
prefect==2.14.0
pytest==8.1.0
